fix: move api clients under clients/ to avoid Playwright route glob collisions
Vite module paths like /src/lib/api/runsClient.ts were matched by e2e **/api/runs** mocks and fulfilled as JSON, white-screening board-token-stats. Relocate domain clients and guard routes with pathname /api/ checks.
