+++ web/src/lib/api/api.ts
@@ -5,14 +5,14 @@ import type { PaginatedResponse } from './apiTypes'
 import { authRedirectPath } from '../composables/useAuth'
 import { req } from './httpCore'
 import type { AuthLoginResponse, AuthMeResponse } from './apiTypes'
-import { projectsClient } from './projectsClient'
-import { pmClient } from './pmClient'
-import { workflowsClient } from './workflowsClient'
-import { runsClient } from './runsClient'
-import { agentsClient } from './agentsClient'
-import { sandboxesClient } from './sandboxesClient'
-import { artifactsClient } from './artifactsClient'
-import { settingsClient } from './settingsClient'
+import { projectsClient } from './clients/projectsClient'
+import { pmClient } from './clients/pmClient'
+import { workflowsClient } from './clients/workflowsClient'
+import { runsClient } from './clients/runsClient'
+import { agentsClient } from './clients/agentsClient'
+import { sandboxesClient } from './clients/sandboxesClient'
+import { artifactsClient } from './clients/artifactsClient'
+import { settingsClient } from './clients/settingsClient'
 
 export function isPaginated<T>(data: T[] | PaginatedResponse<T>): data is PaginatedResponse<T> {
   return data != null && typeof data === 'object' && !Array.isArray(data) && 'items' in data
